:sparkles: feat: add logging for halt event lifecycle status
